Only resave meteor installation in deployment cache if version has changed

+++ bin/meteor.sh
@@ -164,8 +164,13 @@ cache_meteor_install() {
   cache_dir=$2
   meteor_install=$3
   info "Caching meteor runtime for future builds"
-  [ -d "$cache_dir/meteor" ] && rm -rf "$cache_dir/meteor"
-  cp -r "$meteor_install" "$cache_dir/meteor"
+
+  # If first build or new version of meteor, caching meteor install
+  if [ ! -e "$cache_dir/meteor-version" -o "$meteor_version" != "$(cat $cache_dir/meteor-version)" ] ; then
+    [ -d "$cache_dir/meteor" ] && rm -rf "$cache_dir/meteor"
+    cp -r "$meteor_install" "$cache_dir/meteor"
+  fi
+
   [ -d "$cache_dir/meteor-local" ] && rm -rf "$cache_dir/meteor-local"
   cp -r "$build_dir/.meteor/local" "$cache_dir/meteor-local"
   echo $meteor_version > "$cache_dir/meteor-version"
